backend/requirements.txt
amqp==5.0.9
anyio==3.4.0
asgiref==3.4.1
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
bs4==0.0.1
cachetools==5.0.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.1
cssselect==1.1.0
cssutils==2.3.0
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.70.1
Flask==2.0.2
h11==0.12.0
httptools==0.2.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
kombu==5.2.3
lxml==4.7.1
MarkupSafe==2.0.1
orjson==3.6.5
outcome==1.1.0
premailer==3.10.0
prompt-toolkit==3.0.24
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
pymongo==4.0.1
pyOpenSSL==21.0.0
python-dotenv==0.19.2
python-multipart==0.0.5
pytz==2021.3
PyYAML==5.4.1
requests==2.27.1
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
starlette==0.16.0
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==4.0.1
ujson==4.3.0
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
vine==5.0.0
watchgod==0.7
wcwidth==0.2.5
websockets==10.1
Werkzeug==2.0.2
wsproto==1.0.0
yagmail==0.14.260
